DHA: properly implement differences of ACTIVE and INIT SENT state
When we are in initiation_sent state, the logic regarding the neccessity
and treatment of the dialogueInfo is different than active state.

However, it is reasonably similar so that we can use the same function
'is_or_active()' to implement the handling of TR-{CONTINUE,END}.ind
in one function, respectively rather than copy+paste code.